chore: ComposeLoggerFactory relocated
Refs: XRDDEV-2854
